feat: Update default values for app port and host
This commit updates the default values for the app port and host in the ExpressServer constructor. If the environment variables `APP_PORT` or `APP_HOST` are not set, the default values will be used instead. The default port is set to 3000 and the default host is set to "localhost". This change ensures that the app can still run even if the environment variables are not provided.